requirements.txt
anthropic
python-dotenv
requests
Pillow